requirements.txt
# --- Deep Learning Core ---
# Pins minimums for H100 (sm_90) and GradScaler support
torch>=2.11.0
torchvision>=0.26.0
torchaudio>=2.11.0
tensorflow>=2.19.0
keras>=3.13.0
tf-keras>=2.14.0
transformers>=4.40.0
huggingface-hub>=1.8.0

# --- Face Detection & Deepfakes ---
# Note: Requires --no-deps install due to metadata conflicts
# facenet-pytorch>=2.6.0

# --- Data Science & Visualization ---
numpy>=1.26.0
pandas>=3.0.0
scipy>=1.17.0
scikit-learn>=1.8.0
matplotlib>=3.10.0
seaborn>=0.13.0

# --- Utilities & CLI ---
tqdm>=4.67.0
rich>=14.3.0
python-dotenv>=1.2.0
pyyaml>=6.0.0
requests>=2.33.0
click>=8.3.0

# --- Deployment ---
Flask>=3.1.0
Werkzeug>=3.1.0